Do not close the application after an action
Until now the application was closed when a user selected an action. Now
we are reloading the list of alerts and show it again. For that the
"Quit Opsgenie" action was renamed to "Cancel". To close the application
a user can still use "Ctrl + C". The error message which always was
shown when the application was closed, is now also removed.